+++ api/apiHandler.go
@@ -75,6 +75,9 @@ func (a *APIHandler) postHandler(c *gin.Context) {
 	case "deleteToken":
 		a.ApiService.DeleteToken(c)
 		a.apiv2.ReloadTokens()
+	case "setTokenEnabled":
+		a.ApiService.SetTokenEnabled(c)
+		a.apiv2.ReloadTokens()
 	case "logoutAllAdmins":
 		a.ApiService.LogoutAllAdmins(c)
 	default: